requirements.txt
django
django-ckeditor
django-debug-toolbar

pyyaml

# solo en heroku
gunicorn
# psycopg2
django-heroku
python-decouple
whitenoise
